requirements.txt
azure-cognitiveservices-speech
openai
asyncio
discord.py[voice]